requirements.txt
alembic==1.13.3
bcrypt==4.2.1
Faker==30.8.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
pycryptodome==3.21.0
python-dotenv==1.0.1
redis==5.2.0
SQLAlchemy==2.0.36